requirements.txt
py-cord
aiosqlite
ezcord
python-dotenv